backend/requirements.txt
Django
djangorestframework
django-rest-auth
django-cors-headers
psycopg2-binary
pandas
scikit-learn
joblib
gunicorn
torch
torchvision
torchaudio
scikit-learn==0.23.2